Pull up a case docket record on the web

Pulls up a OSCN or ODCR case docket record in the default web browser. This is useful for investigating individual cases for the ways in which items of interest are recorded in online court records.

odcr_show(court, casetype, year, case_seq)

Arguments

case_seq
Sequential case numbers to scrape (i.e., the "45" in "CF-2015-45"), for example, 1 or 1:7000

courts
A character vector of the courts to scrape, for example, "TULSA" or c("TULSA", "ROGERS")

casetypes
A character vector of the case types to scrape, for example, "CF" or c("CF", "CM")

years
A numeric vector of years to scrape, for example, 2016 or 2015:2017

Examples

if (FALSE) {
odcr_show("TULSA", "CF", 2020, 1)
}