Classify hand-entered charges in court and jail data

Uses string detection to standardize dirty administrative records of criminal offenses into analyzable form. This is a work in progress, and only includes a few charges right now: burglary, PWID, and drug trafficking. As we continue to work on new charges, it will be updated with new definitions.

ojo_classify_ct(ct_desc)

Arguments

ct_desc
A string that contains a charge description

Value

The standardized charge

Details

Run `ojo_ct_list` to see the list of charges currently supported by the `ojo_classify_ct()` function.

Examples

ojo_classify_ct("SECOND DEGREE BURGLARY")

#> [1] "BURGLARY II"

ojo_ct_list

#> [1] "BURGLARY I"                            
#> [2] "BURGLARY II"                           
#> [3] "BURGLARY III"                          
#> [4] "POSS BURGLARY TOOLS"                   
#> [5] "BURGLARY (UNSPECIFIED)"                
#> [6] "PWID"                                  
#> [7] "DRUG TRAFFICKING"                      
#> [8] "LARCENY OF MERCHANDISE FROM A RETAILER"