OJO Auth

Configure credentials for the Open Justice Oklahoma database

Usage

ojo_auth(
  host,
  port,
  username,
  password,
  ...,
  .admin = F,
  .overwrite = T,
  .install = T
)

Arguments

host

The host name of the database server

port

The port number of the database server

username

The username to use to connect to the database

password

The password to use to connect to the database

...

Placeholder for additional arguments

.admin

A logical value indicating whether to connect to the database as an administrator

.overwrite

A logical value indicating whether to overwrite the existing .Renviron file

.install

A logical value indicating whether to install the database connection or use it only for the current session

Value

Nothing

Details

Assists the user in populating a .Renviron file with the necessary environment variables to connect to the Open Justice Oklahoma database.

Side Effects

The first time this function is run, it will prompt the user for a username, password, and host name. It will then store these credentials in the user's .Renviron file. If the .Renviron file already exists, it will be backed up and the new credentials will be appended to the end of the file. If the .Renviron file does not exist, it will be created and the credentials will be stored there.

Examples

if (FALSE) {
ojo_auth()
}